Server/requirements.txt
firebase_admin==6.2.0
Flask==3.0.0
Flask_Cors==4.0.0
psycopg2==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1
